Replace special-cased timed events by callbacks
This necessitates some changes:

* Extending the callbacks with an "xpack" parameter that was available
  to the special-cased functions for sending data out from a handler.

* Making synchronisation between execution of the callback and deleting
  the event optional, to not introduce (unnecessary) blocking where it
  wasn't there before.

* Changing the closure argument to a blob that is copied into the event,
  rather than an arbitrary pointer.  The problem is a need to avoid
  races with the garbage collector which makes it impossible to store
  the pointer to an entity without introducing additional stages in the
  deletion process.  This in turn forces an additional indirection in
  the "old" callbacks.

* The internal functions heavily relied on the availability of a "gv"
  pointer, this could easily be emulated via the closure argument, but it
  seems to make more sense to pass it as an argument.

This allows moving a lot of handler code to more reasonable places,
which in turn made "ddsi_discovery" explode too messy to survive without
splitting it.
