+++ src/core/ddsc/tests/deadline.c
@@ -446,9 +446,11 @@ CU_Theory((int32_t n_inst, uint8_t unreg_nth, uint8_t dispose_nth), ddsc_deadlin
 //deadline callback function, this function's purpose is to delay the monitor thread such that while instance's
 //deadline may expire, the event thread is blocked by this function, and updates to instances are "queued" if they happen
 //during this block. these queued updates happen the next time the expiration callbacks fire
-static void cb (struct ddsi_xevent *xev, void *ptr, ddsrt_mtime_t tm)
+static void cb (struct ddsi_domaingv *gv, struct ddsi_xevent *xev, struct ddsi_xpack *xp, void *ptr, ddsrt_mtime_t tm)
 {
+  (void) gv;
   (void) xev;
+  (void) xp;
   (void) ptr;
   (void) tm;
   dds_sleepfor(DEADLINE);
@@ -544,11 +546,11 @@ CU_Test(ddsc_deadline, update)
   }
 
   struct ddsi_domaingv *gvptr = get_domaingv (wr);
-  struct ddsi_xevent *xev = ddsi_qxev_callback(
+  struct ddsi_xevent *xev = ddsi_qxev_callback (
     gvptr->xevents,
     ddsrt_mtime_add_duration(ddsrt_time_monotonic(), (dds_duration_t)(0.5*DEADLINE)),
     cb,
-    NULL);  //this should sleep the thread that updates the statuses from 0.5*DEADLINE to 1.5*DEADLINE
+    NULL, 0, true);  //this should sleep the thread that updates the statuses from 0.5*DEADLINE to 1.5*DEADLINE
   CU_ASSERT_FATAL(xev != NULL);
 
   Space_Type1 msg1 = { 1, 0, 0 },
@@ -599,7 +601,7 @@ CU_Test(ddsc_deadline, update)
   //msg1 should have expired 3 times, msg2 should have expired 2 times
   check_statuses(wr, rd, expired_1, expired_2, tw1, tw2, ih1, ih2);
 
-  ddsi_delete_xevent_callback(xev);
+  ddsi_delete_xevent(xev);
   dds_delete(pp);
 }
