+++ src/core/ddsc/tests/nwpart.c
@@ -27,6 +27,7 @@
 #include "ddsi__misc.h"
 #include "ddsi__addrset.h"
 #include "ddsi__discovery.h"
+#include "ddsi__discovery_endpoint.h"
 #include "ddsi__plist.h"
 #include "ddsi__tran.h"
