+++ src/core/ddsi/CMakeLists.txt
@@ -64,6 +64,10 @@ set(srcs_ddsi
   ddsi_guid.c
   ddsi_bswap.c
   ddsi_discovery.c
+  ddsi_discovery_addrset.c
+  ddsi_discovery_spdp.c
+  ddsi_discovery_sedp.c
+  ddsi_discovery_endpoint.c
   ddsi_debmon.c
   ddsi_init.c
   ddsi_lat_estim.c
@@ -188,6 +192,10 @@ set(hdrs_private_ddsi
   ddsi__bitset.h
   ddsi__bswap.h
   ddsi__discovery.h
+  ddsi__discovery_addrset.h
+  ddsi__discovery_spdp.h
+  ddsi__discovery_sedp.h
+  ddsi__discovery_endpoint.h
   ddsi__debmon.h
   ddsi__hbcontrol.h
   ddsi__inverse_uint32_set.h
@@ -220,6 +228,14 @@ if(ENABLE_DEADLINE_MISSED)
   )
   list(APPEND hdrs_ddsi
     ddsi_deadline.h
+  )
+endif()
+if(ENABLE_TOPIC_DISCOVERY)
+  list(APPEND srcs_ddsi
+    ddsi_discovery_topic.c
+  )
+  list(APPEND hdrs_private_ddsi
+    ddsi__discovery_topic.h
   )
 endif()
 if(ENABLE_TYPE_DISCOVERY)
