+++ src/core/ddsi/src/ddsi_transmit.c
@@ -1361,3 +1361,12 @@ int ddsi_write_sample_nogc_notk (struct ddsi_thread_state * const thrst, struct 
   ddsi_tkmap_instance_unref (wr->e.gv->m_tkmap, tk);
   return res;
 }
+
+int ddsi_write_and_fini_plist (struct ddsi_writer *wr, ddsi_plist_t *ps, bool alive)
+{
+  struct ddsi_serdata *serdata = ddsi_serdata_from_sample (wr->type, alive ? SDK_DATA : SDK_KEY, ps);
+  ddsi_plist_fini (ps);
+  serdata->statusinfo = alive ? 0 : (DDSI_STATUSINFO_DISPOSE | DDSI_STATUSINFO_UNREGISTER);
+  serdata->timestamp = ddsrt_time_wallclock ();
+  return ddsi_write_sample_nogc_notk (ddsi_lookup_thread_state (), NULL, wr, serdata);
+}
